Improved the execution time thanks to converting BigDecimal to int
Former-commit-id: 9c997205c29015ceb978cfac6a3e3c559c17104f [formerly eeb65a3417f2db314e95f0779d5403511dcf46ae]
Former-commit-id: 9320c4e90da1eef33e6b51849476aa2592f44bfd

+++ vaccine-management-center/src/main/java/pl/wysockif/optimizer/items/connections/Connection.java
@@ -9,9 +9,9 @@ public class Connection {
     private final Producer producer;
     private final Pharmacy pharmacy;
     private final int maxNumberOVaccines;
-    private final double price;
+    private final int price;
 
-    public Connection(Producer producer, Pharmacy pharmacy, int maxNumberOVaccines, double price) {
+    public Connection(Producer producer, Pharmacy pharmacy, int maxNumberOVaccines, int price) {
         this.producer = producer;
         this.pharmacy = pharmacy;
         this.maxNumberOVaccines = maxNumberOVaccines;
@@ -44,7 +44,7 @@ public class Connection {
         return maxNumberOVaccines;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
